operaciones.py separado en tres archivos colocados en managers/
