+++ gauss_bot/managers/mats_manager.py
@@ -1,29 +1,21 @@
 from fractions import Fraction
-from typing import Tuple, List, Dict, overload
 from copy import deepcopy
+from typing import Tuple, List, overload
 
-from .clases.matriz import Matriz
-from .clases.vector import Vector
-from .clases.sistema_ecuaciones import SistemaEcuaciones
-from .utils import limpiar_pantalla, match_input
-
-DictMatrices = Dict[str, Matriz]
-DictVectores = Dict[str, Vector]
-
-# TODO: mejorar .procesar_operacion() en MatricesManager y VectoresManager
-# TODO implementer producto matriz-vector en ambos menus
+from gauss_bot.clases.matriz import Matriz, DictMatrices
+from gauss_bot.clases.sistema_ecuaciones import SistemaEcuaciones
+from gauss_bot.utils import limpiar_pantalla, match_input
 
 class MatricesManager:
-    def __init__(self, vec_manager: "VectoresManager", mats_ingresadas: DictMatrices = {}) -> None:
-        self.vec_manager = vec_manager
+    def __init__(self, mats_ingresadas: DictMatrices = {}) -> None:
         self.mats_ingresadas = mats_ingresadas
 
     def menu_matrices(self) -> None:
         while True:
             limpiar_pantalla()
-            print("\n###########################")
+            print("\n###############################")
             print("### Operaciones de Matrices ###")
-            print("###########################")
+            print("###############################")
             print("\n1. Agregar una matriz")
             print("2. Mostrar matrices\n")
             print("3. Resolver sistema de ecuaciones")
@@ -52,7 +44,7 @@ class MatricesManager:
             case 1:
                 self.agregar_matriz()
             case 2:
-                self._imprimir_matrices(necesita_aumentada = -1)
+                self._mostrar_matrices(necesita_aumentada = -1)
             case 3:
                 self.procesar_operacion("se")
             case 4:
@@ -166,10 +158,9 @@ class MatricesManager:
         * "s": sumar matrices
         * "r": restar matrices
         * "t": transponer matriz
-        * "mv": multiplicar matriz por vector
         """
 
-        if operacion not in ("se", "s", "r", "me", "m", "t", "mv"):
+        if operacion not in ("se", "s", "r", "me", "m", "t"):
             return None
 
         if operacion in ("s", "r", "m"):
@@ -206,9 +197,6 @@ class MatricesManager:
                     if nombre_mat_modded not in self.mats_ingresadas.keys():
                         self.mats_ingresadas[nombre_mat_modded] = mat_copia_modded
                     return (nombre_mat, mat_copia_modded)
-        
-        elif operacion == "mv":
-            ...  # TODO: producto matriz-vector
 
     @overload
     def seleccionar(self, operacion: str) -> str: ...
@@ -255,7 +243,7 @@ class MatricesManager:
     def _get_input(self, mensaje: str, operacion: str | None) -> str:
         limpiar_pantalla()
         necesita_aumentada = operacion == "se"
-        self._imprimir_matrices(necesita_aumentada)
+        self._mostrar_matrices(necesita_aumentada)
         return input(mensaje).strip()
 
     def _validate_input_mat(self, input_mat: str, operacion: str) -> None:
@@ -277,7 +265,7 @@ class MatricesManager:
             return False
         return True
 
-    def _imprimir_matrices(self, necesita_aumentada: int) -> None:
+    def _mostrar_matrices(self, necesita_aumentada: int) -> None:
         """
         ? necesita_aumentada:
         *  1: necesita que la matriz sea aumentada
@@ -297,253 +285,5 @@ class MatricesManager:
             print(mat)
         print("---------------------------------------------")
 
+        input("\nPresione cualquier tecla para regresar al menú de matrices...")
         return None
-
-
-class VectoresManager:
-    def __init__(self, mat_manager: MatricesManager, vecs_ingresados: DictVectores = {}) -> None:
-        self.mat_manager = mat_manager
-        self.vecs_ingresados = vecs_ingresados
-
-    def menu_vectores(self) -> None:
-        while True:
-            limpiar_pantalla()
-            print("\n###########################")
-            print("### Operaciones de Vectores ###")
-            print("###########################")
-            print("\n1. Agregar un vector")
-            print("2. Mostrar vectores")
-            print("3. Sumar vectores")
-            print("4. Restar vectores")
-            print("5. Multiplicar vector por escalar")
-            print("6. Multiplicar vectores")
-            print("7. Multiplicar matriz por vector")
-            print("8. Regresar al menú principal")
-            try:
-                option = int(input("\nSeleccione una opción: ").strip())
-                if option < 1 or option > 8:
-                    raise ValueError
-            except ValueError:
-                input("\nError: Ingrese una opción válida!")
-                continue
-            if option == 8:
-                input("Regresando a menú principal...")
-                break
-            self.procesar_menu(option)
-        return None
-
-    def procesar_menu(self, opcion: int) -> None:
-        match opcion:
-            case 1:
-                self.agregar_vector()
-            case 2:
-                self._imprimir_vectores()
-            case 3:
-                self.procesar_operacion("s")
-            case 4:
-                self.procesar_operacion("r")
-            case 5:
-                self.procesar_operacion("ve")
-            case 6:
-                self.procesar_operacion("m")
-            case 7:
-                ...  # TODO: producto matriz-vector
-        return None
-
-    def agregar_vector(self) -> None:
-        try:
-            limpiar_pantalla()
-            nombre = input("\nIngrese el nombre del vector (una letra minúscula): ").strip().lower()
-            if not nombre.isalpha() or len(nombre) != 1:
-                raise NameError("Error: Ingrese solamente una letra minúscula!")
-            if nombre in self.vecs_ingresados:
-                raise KeyError(f"Error: Ya hay un vector con el nombre {nombre}!")
-
-            longitud = int(input("¿Cuántas dimensiones tendrá el vector? ").strip())
-
-        except NameError as n:
-            input(n)
-            return self.agregar_vector()
-        except KeyError as k:
-            input(k)
-            return self.agregar_vector()
-        except ValueError:
-            input("Error: Ingrese un número entero!")
-            return self.agregar_vector()
-
-        self.vecs_ingresados[nombre] = self.pedir_vector(longitud)
-        print("\nMatriz agregada exitosamente!")
-        ingresar_otro = match_input("¿Desea ingresar otro vector? (s/n) ")
-        if ingresar_otro:
-            return self.agregar_vector()
-        elif not ingresar_otro:
-            input("Regresando al menú de vectores...")
-        else:
-            input("Opción inválida! Regresando al menú de vectores...")
-        return None
-
-    def pedir_vector(self, longitud: int) -> Vector:
-        try:
-            componentes = [
-                Fraction(x).limit_denominator(100)
-                for x in input("Ingrese los componentes (separados por comas): ")
-                .strip()
-                .split(",")
-            ]
-
-            if len(componentes) != longitud:
-                raise TypeError("Error: El vector no tiene las dimensiones correctas!")
-
-        except TypeError as t:
-            input(t)
-            return self.pedir_vector(longitud)
-        except ValueError:
-            input("Error: Ingrese un número real!")
-            return self.pedir_vector(longitud)
-        except ZeroDivisionError:
-            input("Error: El denominador no puede ser cero!")
-            return self.pedir_vector(longitud)
-
-        return Vector(componentes)
-
-    def procesar_operacion(self, operacion: str):
-        """
-        ? operacion:
-        * "s": sumar vectores
-        * "r": restar vectores
-        * "ve": multiplicar vector por escalar
-        * "m": multiplicar vectores
-        * "mv": multiplicar matriz por vector
-        """
-
-        if operacion not in ("s", "r", "ve", "m", "mv"):
-            return None
-
-        if operacion in ("s", "r", "m"):
-            nombres_vecs = self.seleccionar(None)
-            vec1, vec2 = self.vecs_ingresados[nombres_vecs[0]], self.vecs_ingresados[nombres_vecs[1]]
-            match operacion:
-                case "s":
-                    return (nombres_vecs, vec1 + vec2)
-                case "r":
-                    return (nombres_vecs, vec1 - vec2)
-                case "m":
-                    return (nombres_vecs, vec1 * vec2)
-
-        elif operacion == "ve":
-            nombre_vec = self.seleccionar(operacion)
-            vec_copia = deepcopy(self.vecs_ingresados[nombre_vec])
-            escalar = Fraction(input("Ingrese el escalar: ").strip())
-            return (nombre_vec, vec_copia * escalar)
-        elif operacion == "mv":
-            ...  # TODO: producto matriz-vector
-
-    @overload
-    def seleccionar(self, operacion: str) -> str: ...
-
-    @overload
-    def seleccionar(self, operacion: None) -> List[str]: ...
-
-    def seleccionar(self, operacion: str | None) -> str | List[str]:
-        if operacion not in ("ve", "mv", None):
-            return "" if operacion else []
-
-        if not self._validar_vecs_ingresados():
-            return "" if operacion else []
-
-        if operacion is not None:
-            mensaje = self._get_mensaje(operacion)
-            input_vec = self._get_input(mensaje, operacion)
-            self._validate_input_vec(input_vec, operacion)
-            return input_vec
-        else:
-            mensaje = "\n¿Cuáles vectores desea seleccionar? (separados por comas) "
-            input_vecs = self._get_input(mensaje, operacion).split(",")
-            self._validate_input_vecs(input_vecs)
-            return input_vecs
-
-    def _get_mensaje(self, operacion: str) -> str:
-        match operacion:
-            case "ve":
-                return "\n¿Cuál vector desea multiplicar por un escalar? "
-            case "mv":
-                return "\n¿Cuál vector desea multiplicar por una matriz? "
-            case _:
-                return ""
-
-    def _get_input(self, mensaje: str, operacion: str | None) -> str:
-        limpiar_pantalla()
-        self._imprimir_vectores()
-        return input(mensaje).strip()
-
-    def _validate_input_vec(self, input_vec: str, operacion: str) -> None:
-        if input_vec not in self.vecs_ingresados:
-            raise KeyError("Error: El vector no existe.")
-
-    def _validate_input_vecs(self, input_vecs: List[str]) -> None:
-        vec = next((vec for vec in input_vecs if vec not in self.vecs_ingresados), None)
-        if vec is not None:
-            raise KeyError(f"Error: El vector {vec} no existe.")
-        elif len(input_vecs) != 2:
-            raise ValueError("Error: Debe seleccionar exactamente dos vectores.")
-
-    def _validar_vecs_ingresados(self) -> bool:
-        if self.vecs_ingresados == {}:
-            input("\nNo hay vectores ingresados!")
-            return False
-        return True
-
-    def _imprimir_vectores(self, es_matricial=False) -> None:
-        if not self._validar_vecs_ingresados():
-            return None
-
-        mensaje = "seleccionados" if es_matricial else "ingresados"
-        print(f"\nVectores {mensaje}:")
-        print("---------------------------------------------")
-        for nombre, vec in self.vecs_ingresados.items():
-            print(f"{nombre}: {vec}")
-        print("---------------------------------------------")
-
-        return None
-
-
-class OpsManager:
-    def __init__(self, mat_manager: MatricesManager, vec_manager: VectoresManager) -> None:
-        self.mat_manager = mat_manager
-        self.vec_manager = vec_manager
-
-    def start_exec_loop(self) -> None:
-        while True:
-            option = self.main_menu()
-            match option:
-                case 1:
-                    self.mat_manager._imprimir_matrices(necesita_aumentada = -1)
-                case 2: ...
-                    # self.vec_manager._imprimir_vectores()
-                case 3:
-                    self.mat_manager.menu_matrices()
-                case 4:
-                    self.vec_manager.menu_vectores()
-                case 5:
-                    input("Cerrando programa...")
-                    break
-        return None
-
-    def main_menu(self) -> int:
-        limpiar_pantalla()
-        print("\n################")
-        print("### GaussBot ###")
-        print("################\n")
-        print("1. Mostrar matrices")
-        print("2. Mostrar vectores")
-        print("3. Operaciones de matrices")
-        print("4. Operaciones de vectores")
-        print("5. Cerrar programa")
-        try:
-            option = int(input("\n=> Seleccione una opción: ").strip())
-            if option < 1 or option > 5:
-                raise ValueError
-        except ValueError:
-            input("\nError: Ingrese una opción válida!")
-            return self.main_menu()
-        return option
