+++ main.py
@@ -1,4 +1,4 @@
-from gauss_bot.operaciones import OpsManager, MatricesManager, VectoresManager
+from gauss_bot.managers.main_manager import OpsManager, MatricesManager, VectoresManager
 
 if __name__ == "__main__":
     manager = OpsManager(MatricesManager(), VectoresManager())
